Fix #1366 announce-stripe-connect error handle endpoint downtime gracefully
